Added Fossil entity: migration, model, templates, controller, routes. Added Fossil in Rock actions. Fixed saving amd removing of minerals. Fixed removing of detail and microscope photos. Refactored getting of links on Rock detail page - getMineralLinkItems. Fixed id type to bigInteger in migrations.

+++ database/migrations/2021_06_25_105657_create_rock__forming_minerals_table.php
@@ -15,8 +15,8 @@ class CreateRockFormingMineralsTable extends Migration
     {
         Schema::create('rock__forming_minerals', function (Blueprint $table) {
             $table->id();
-            $table->integer('rock_id');
-            $table->integer('mineral_id');
+            $table->bigInteger('rock_id');
+            $table->bigInteger('mineral_id');
             $table->timestamps();
         });
     }
