+++ database/migrations/2021_06_25_110141_create_rock__second_minerals_table.php
@@ -15,8 +15,8 @@ class CreateRockSecondMineralsTable extends Migration
     {
         Schema::create('rock__second_minerals', function (Blueprint $table) {
             $table->id();
-            $table->integer('rock_id');
-            $table->integer('mineral_id');
+            $table->bigInteger('rock_id');
+            $table->bigInteger('mineral_id');
             $table->timestamps();
         });
     }
